cpp-plugin: optimized DWARF reader
Added a sorted vector instead of dwarf::line_entry used before. Allows
for fast lookup of dwarf infos. This was needed to be able to load rust
binaries which have a lot of pc_range in their DWARF info.

We are now able to load big debug binaries within seconds.