pp: detect loops and report stats about it + memory flamegraph
we now detect loops during execution (detection is made everytime a new
transition between two basic blocks is created), and keep a loop and
call stack (intermixed) to represent current context of execution. This
stack is maintained per thread.

For every element on this context, we associate some statistics
(instructions, memory read/written, ...). The stack allows us to record
statistics for each context (loop or symbol), and for each cumulated
context (when this symbol or loop is on the stack). This allows to have
a useful view on the program performance.

cpu_flame_graph_profiler plugin was removed, since its functionality is
not needed anymore.

added flamegraph for memory read/write in more than cpu instructions.

IMPORTANT Note: due to algorithm we use to detect loop, first iteration
of an inner loop will not be reported as being nested (since we don't
know yet there is an higher level loop).